# PERF: improves performance in SeriesGroupBy.count