@jreback 2015-09-02
@behzadnouri pls rebase and make the changes as above