@behzadnouri 2015-09-05
@jreback the added test will fail if i change it what u suggest.