@jreback 2015-09-05
https://github.com/pydata/pandas/issues/443 would fix this, e.g. we need an option like:

``s.count(level=1, skipna=False)``