@jreback 2015-09-05
merged via https://github.com/pydata/pandas/commit/33723f9577d451c20db68b4d10ed90e81f54e03e